Create recurrence_relation.c
Complete the below function to implement a recurrence relation T is defined on n >= 0 and is given as

T(n) = T(n-1) + 3*n

with the base case T(0) = 0.

You will be given one integer k and you have to write a program to find out T(k). The program must implement T( ) recursively.