Create Flood_fill_algo.c
In image editing softwares like Paint, when we take the brush to a pixel and click, the color of the region of that pixel is replaced with a new selected color. The problem statement to do this task is as below: 

Given a 2-Dimensional screen, location of a pixel in the screen (x, y coordinates of that pixel) and a color, your task is to replace color of the given pixel and all adjacent same colored pixels with the given color.

Note: Maximum screen size in this question is taken as 50*50.

Input Format

First Line will contain two integers M and N denoting the size of matrix.
Second line contains M*N integers as matrix elements row-wise separated by space.
Third line contains 3 integers denoting xCord, YCord, newColor for the pixel.
Output Format

Print the new matrix as M*N integers row-wise separated by space.
Sample Input

4 4
1 1 1 1 1 2 2 1 1 1 2 1 1 2 2 2
1 2 3
Sample Output

1 1 1 1 1 3 3 1 1 1 3 1 1 3 3 3
Explanation

pixel at location 1,2 has color = 2 already and all adjacent pixels will now be replaced with color = 3.